Optimise Pop()
The performance of the Push()-Pop() calling pattern decreased since
changing from Poll() to Pop(), and including item removal in the latter
(as is typical in queues).

Rework the Pop() function such that the return is not copied into a
temporary. Also create a local rIdx copy to avoid continuous reads of
the value in the queue structure's memory.

Benchmark results now compare better with the Offer()-Peek()-Advance()
pattern.